==> Unsupervised Learning/2. Predicting Fish Categories.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

fish_df = pd.read_csv('https://assets.datacamp.com/production/repositories/655/datasets/fee715f8cf2e7aad9308462fea5a26b791eb96c4/fish.csv',header=None)

fish_df.head()

fish_df.info()

samples = fish_df.iloc[:,1:6]

species = fish_df.iloc[:,0]

from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import KMeans

kmeans = KMeans(n_clusters=4)
kmeans.fit(samples)
labels = kmeans.predict(samples)

cross_df = pd.DataFrame({'labels':labels,'species':species})

cross_tab = pd.crosstab(cross_df['labels'],cross_df['species'])

print(cross_tab)

scaler = StandardScaler()
kmeans = KMeans(n_clusters=4)
pipeline = make_pipeline(scaler,kmeans)

pipeline.fit(samples)
labels = pipeline.predict(samples)

cross_df = pd.DataFrame({'labels':labels,'species':species})

cross_tab = pd.crosstab(cross_df['labels'],cross_df['species'])

print(cross_tab)
