==> Unsupervised Learning/7. Spotify Recommendations -- Data Prep.ipynb <==
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import re
import numpy as np
import pandas as pd

client_credentials_manager = SpotifyClientCredentials(client_id='cf5e39da3b50438493846ee960df886d',client_secret='23e7cc1dcb4f4836bc594667c6e2cef3')
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

def track_features(trackId):
    features = sp.audio_features(trackId)[0]
    artist = sp.track(trackId)['artists'][0]['id']
    artist_pop = sp.artist(artist)['popularity']
    artist_genres =  sp.artist(artist)['genres']
    
    track_pop = sp.track(trackId)['popularity']
    
    features['artist_pop'] = artist_pop
    if artist_genres:
        features['genres'] = " ".join([re.sub(' ','_',i) for i in artist_genres])
    else:
        features['genres'] = 'unknown'
        
    features['track_pop'] = track_pop
    
    return features

playlist = sp.playlist('37i9dQZF1DXdOEFt9ZX0dh')

playlist_df = pd.DataFrame(columns=['artist_name','track_uri','artist_uri','track_name','album_uri','duration_ms','album_name'])

for i in range(0,len(playlist['tracks']['items'])):
    attribute_list=[]
    attribute_list.append(playlist['tracks']['items'][i]['track']['artists'][0]['name'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['uri'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['artists'][0]['uri'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['name'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['album']['uri'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['duration_ms'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['album']['name'])
    playlist_df.loc[len(playlist_df)] = attribute_list

playlist_df['track_uri'] = playlist_df['track_uri'].apply(lambda x:x[x.rfind(':')+1:])

playlist_df.head()

key_columns=[]
for key in track_features(playlist_df['track_uri'][0]):
    key_columns.append(key)

track_features_df = pd.DataFrame(columns=key_columns)

featureList = []
for i in range(0,len(playlist_df)):
#     print(i)
    featureList.append(track_features(playlist_df['track_uri'][i]))

track_features_df = pd.DataFrame(featureList)

track_features_df.head()

music_df = pd.merge(playlist_df,track_features_df,left_on='track_uri',right_on = 'id')

music_df = music_df.drop_duplicates()

music_df.info()

music_df.describe()

music_df.to_csv('C:\\Users\\Akhil_S_Bhat\\Documents\\AIML\Machine_Learning\\Unsupervised Learning\\Data\\3_Spotify_Data\\spotify_test_data.csv')
