==> Unsupervised Learning/8. Spotify Recommendations -- Feature Extraction.ipynb <==
import numpy as np
import pandas as pd
import re

from textblob import TextBlob

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import MinMaxScaler

spotify_data = pd.read_csv('C:\\Users\\Akhil_S_Bhat\\Documents\\AIML\Machine_Learning\\Unsupervised Learning\\Data\\3_Spotify_Data\\spotify_extracted_data.csv')

spotify_data.head()

spotify_data['genres'] = spotify_data['genres'].apply(lambda x:x.split(' '))

columns_to_use=['artist_name','id','track_name','artist_uri','album_uri',
                'duration_ms_x','album_name','danceability','energy','key', 'loudness',
                'mode','speechiness','acousticness','instrumentalness','liveness','valence',
                'tempo','genres', 'track_pop','artist_pop']

spotify_data = spotify_data[columns_to_use]

def getSubjectivity(text):
    return TextBlob(text).sentiment.subjectivity

def getPolarity(text):
    return TextBlob(text).sentiment.polarity

def getAnalysis(score, task='polarity'):
    if task == 'subjectivity':
        if score < 1/3:
            return 'low'
        elif score > 1/3:
            return 'high'
        else:
            return 'medium'
    else:
        if score < 0 :
            return 'Negative'
        elif score == 0:
            return 'Neutral'
        else:
            return 'Positive'
        
def sentiment_analysis(df,text_col):
    df['subjectivity'] = df[text_col].apply(getSubjectivity).apply(lambda x:getAnalysis(x,'subjectivity'))
    df['polarity'] = df[text_col].apply(getPolarity).apply(getAnalysis)
    return df

sentiment = sentiment_analysis(spotify_data, "track_name")

subjectivity_columns = pd.get_dummies(spotify_data['subjectivity'])

subjectivity_columns = subjectivity_columns.rename(columns={'high':'subjectivity_high','low':'subjectivity_low'})

spotify_data = pd.concat([spotify_data,subjectivity_columns],axis=1)

spotify_data = spotify_data.drop(['subjectivity'],axis=1)

polarity_columns = pd.get_dummies(spotify_data['polarity']) # 'Negative', 'Neutral', 'Positive'

polarity_columns = polarity_columns.rename(columns={'Negative':'polarity_Negative','Neutral':'polarity_Neutral','Positive':'polarity_Positive'})

spotify_data = pd.concat([spotify_data,polarity_columns],axis=1)

spotify_data = spotify_data.drop(['polarity'],axis=1)

key_columns = pd.get_dummies(spotify_data['key'])

key_columns = key_columns.rename(columns={0:'key0',1:'key1',2:'key2',3:'key3',4:'key4',5:'key5',6:'key6',7:'key7',8:'key8',9:'key9',10:'key10',11:'key11'})

spotify_data = pd.concat([spotify_data,key_columns],axis=1)

spotify_data = spotify_data.drop(['key'],axis=1)

mode_columns = pd.get_dummies(spotify_data['mode'])

mode_columns = mode_columns.rename(columns={0:'mode0',1:'mode1'})

spotify_data = pd.concat([spotify_data,mode_columns],axis=1)

spotify_data = spotify_data.drop(['mode'],axis=1)

spotify_data.head()

tfidf = TfidfVectorizer()
tfidf_matrix = tfidf.fit_transform(spotify_data['genres'].apply(lambda x: " ".join(x)))

genre_df = pd.DataFrame(tfidf_matrix.toarray())
genre_df.columns = tfidf.get_feature_names()

pop = spotify_data[['artist_pop','track_pop']].reset_index(drop=True)
scaler = MinMaxScaler()
pop_scaled = pd.DataFrame(scaler.fit_transform(pop),columns=pop.columns)
pop_scaled.head()

float_columns = spotify_data.dtypes[spotify_data.dtypes=='float64'].index.values
float_data = spotify_data[float_columns].reset_index(drop=True)
float_scaled = pd.DataFrame(scaler.fit_transform(float_data),columns=float_columns)
float_scaled.head()

spotify_data=spotify_data.drop(float_scaled,axis=1)

final_df = pd.concat([spotify_data,genre_df,pop_scaled,float_scaled],axis=1)

final_df.head()

final_df.columns.values

final_df.columns.values

final_df.to_csv('C:\\Users\\Akhil_S_Bhat\\Documents\\AIML\\Machine_Learning\\Unsupervised Learning\\Data\\3_Spotify_Data\\complete_feature.csv')
