==> Unsupervised Learning/9. Spotify Recommendations -- Modeling.ipynb <==
import numpy as np
import pandas as pd

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

from sklearn.metrics.pairwise import cosine_similarity

playlistDF_test = pd.read_csv('C:\\Users\\Akhil_S_Bhat\\Documents\\AIML\\Machine_Learning\\Unsupervised Learning\\Data\\3_Spotify_Data\\spotify_test_data.csv')

complete_feature_set = pd.read_csv('C:\\Users\\Akhil_S_Bhat\\Documents\\AIML\\Machine_Learning\\Unsupervised Learning\\Data\\3_Spotify_Data\\complete_feature.csv')

complete_feature_set_playlist = complete_feature_set[complete_feature_set['id'].isin(playlistDF_test['id'].values)]

complete_feature_set_nonplaylist = complete_feature_set[~complete_feature_set['id'].isin(playlistDF_test['id'].values)]

complete_feature_set_playlist_vector = complete_feature_set_playlist.drop(columns=['Unnamed: 0','id','artist_name','track_name','artist_uri','album_uri','album_name','genres']).sum(axis=0)

client_credentials_manager = SpotifyClientCredentials(client_id='cf5e39da3b50438493846ee960df886d',client_secret='23e7cc1dcb4f4836bc594667c6e2cef3')
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

playlist = sp.playlist('5j14kA383XcoO33Px4HFPL')

playlist_df = pd.DataFrame(columns=['artist_name','track_uri','artist_uri','track_name','album_uri','duration_ms','album_name'])

for i in range(0,len(playlist['tracks']['items'])):
    attribute_list=[]
    attribute_list.append(playlist['tracks']['items'][i]['track']['artists'][0]['name'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['uri'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['artists'][0]['uri'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['name'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['album']['uri'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['duration_ms'])
    attribute_list.append(playlist['tracks']['items'][i]['track']['album']['name'])
    playlist_df.loc[len(playlist_df)] = attribute_list

playlist_df['track_uri'] = playlist_df['track_uri'].apply(lambda x:x[x.rfind(':')+1:])

playlist_df.head()

playlist_df.shape

non_playlist_df = playlist_df[playlist_df['track_uri'].isin(complete_feature_set_nonplaylist['id'].values)]

len(complete_feature_set_playlist_vector.values.reshape(1, -1))

complete_feature_set_nonplaylist = complete_feature_set_nonplaylist.drop(columns=['Unnamed: 0','id','artist_name','track_name','artist_uri','album_uri','album_name','genres'],axis=1)

non_playlist_df = non_playlist_df.drop_duplicates()

non_playlist_df.shape

non_playlist_df['sim'] = cosine_similarity(complete_feature_set_nonplaylist.values, complete_feature_set_playlist_vector.values.reshape(1, -1))[:,0]

non_playlist_df_top_10 = non_playlist_df.sort_values('sim',ascending = False).head(10)

non_playlist_df_top_10[['artist_name','track_name']]
